explore & analyze seperated out into individual tabs. Data air table replaced with ref table
